Clean up file names, move components into directories
