--- Feed.js
@@ -1,12 +1,0 @@
-import React, { Component } from "react";
-import { View, StyleSheet, Text, Button } from "react-native";
-
-export default class Feed extends Component {
-  render() {
-    return (
-      <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
-        <Text>Feed</Text>
-      </View>
-    );
-  }
-}
