--- Settings.js
@@ -1,18 +1,0 @@
-import React, { Component } from "react";
-import { View, StyleSheet, Text, Button } from "react-native";
-
-export default class Settings extends Component {
-  render() {
-    return (
-      <View
-        style={{
-          flex: 1,
-          alignItems: "center",
-          justifyContent: "center"
-        }}
-      >
-        <Text>Settings</Text>
-      </View>
-    );
-  }
-}
