+++ components/WelcomeScreen/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  login: {
+    flex: 1, 
+    alignItems: "center", 
+    justifyContent: "center" 
+  }
+});
+
+export default styles;
